+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add position in frame to attendee updates
 - Add stale-issue bot configuration
 - Add simulcast support and provides new uplink downlink policies
-
+- Add MultiLogger to support logging to multiple Logger instances
 
 ### Changed
 - Use GET instead of POST to obtain TURN credentials